requirements.txt
# Core dependencies for RealtimeVoiceAPI
websockets>=11.0.0

aiohttp>=3.8.0

# Environment variable management
python-dotenv>=1.0.0

# Audio processing
numpy>=1.21.0
pydub>=0.25.0

# Optional dependencies for enhanced functionality
scipy>=1.9.0  # For advanced audio processing
matplotlib>=3.5.0  # For audio visualization (optional)

# pyaudio

sounddevice
soundfile

# Development dependencies (uncomment if needed)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=22.0.0
# mypy>=1.0.0